refactor: PhoneNumberVisualTransformation 위치 변경 (core:common -> core:ui)
- `PhoneNumberInputTransformation` 및 `PhoneNumberVisualTransformation`을 UI 관련 유틸리티로 판단하여 `core:common` 패키지에서 `core:ui` 패키지로 이동
- 패키지 위치 변경에 따라 `AddAfternoteEditorReceiverDialog.kt` 내 관련 임포트 경로 업데이트
